fix(subprocess): action follow-up review findings on portability and tests
- stream_pipes: probe pipe capacity via a guarded _probe_pipe_capacity that
  falls back to a conservative default when fcntl or F_GETPIPE_SZ is absent,
  so drain_blocking_payload_size callers need no platform guard.
- _subprocess_execution: reuse _drain_stream_consumers in the stdin-writer
  failure handler instead of a manual cancel+gather, matching the timeout and
  cancellation branches.
- test_subprocess_timeout: give the drain hypothesis test a fixed settling
  window so slow consumers stay pending and exercise cancellation, expecting
  None past the window.
- test_observe: extract the early-close coordination cluster into
  test_observe_stdin_early_close.py, bringing test_observe.py under the
  400-line cap; the child now waits for the go marker with a finite deadline
  and exits non-zero rather than spinning forever.
- developers-guide: correct the streaming-timeout flow description and
  document the shared drain_blocking_payload_size helper and its reuse policy.
- Regenerate the maturin wheel-manifest snapshot for the new test module.
